refactor: remove call_llm, move pdf functions to a separate module
